Move Entities to Domain package
This commit moves the `entities` package to be under the `domain` module as it was missing from the architecture.
