+++ fetcher/src/main/kotlin/com/xkcddatahub/fetcher/adapters/output/database/postgres/WebComicPostgresAdapter.kt
@@ -4,7 +4,7 @@ import com.xkcddatahub.fetcher.adapters.output.database.postgres.mappers.Databas
 import com.xkcddatahub.fetcher.adapters.output.database.postgres.table.WebComicsTable
 import com.xkcddatahub.fetcher.application.ports.WebComicsPort
 import com.xkcddatahub.fetcher.bootstrap.DatabaseFactory.Companion.dbQuery
-import com.xkcddatahub.fetcher.entity.WebComics
+import com.xkcddatahub.fetcher.domain.entity.WebComics
 import org.jetbrains.exposed.sql.insert
 import org.jetbrains.exposed.sql.max
