+++ fetcher/src/main/kotlin/com/xkcddatahub/fetcher/application/ports/WebComicsPort.kt
@@ -1,6 +1,6 @@
 package com.xkcddatahub.fetcher.application.ports
 
-import com.xkcddatahub.fetcher.entity.WebComics
+import com.xkcddatahub.fetcher.domain.entity.WebComics
 
 interface WebComicsPort {
     suspend fun save(comics: WebComics): Boolean
