+++ fetcher/src/main/kotlin/com/xkcddatahub/fetcher/application/ports/XkcdClientPort.kt
@@ -1,6 +1,6 @@
 package com.xkcddatahub.fetcher.application.ports
 
-import com.xkcddatahub.fetcher.entity.WebComics
+import com.xkcddatahub.fetcher.domain.entity.WebComics
 
 interface XkcdClientPort {
     suspend fun getLatestComicId(): Int
