+++ fetcher/src/test/kotlin/com/xkcddatahub/fetcher/output/database/postgres/mappers/DatabaseWebComicsMapperTest.kt
@@ -2,7 +2,7 @@ package com.xkcddatahub.fetcher.output.database.postgres.mappers
 
 import com.xkcddatahub.fetcher.adapters.output.database.postgres.data.DatabaseWebComics
 import com.xkcddatahub.fetcher.adapters.output.database.postgres.mappers.DatabaseWebComicsMapper.toData
-import com.xkcddatahub.fetcher.entity.WebComics
+import com.xkcddatahub.fetcher.domain.entity.WebComics
 import io.kotest.matchers.shouldBe
 import kotlin.test.Test
