+++ fetcher/src/test/kotlin/com/xkcddatahub/fetcher/output/http/rest/xkcd/mappers/ApiWebComicsMapperTest.kt
@@ -2,7 +2,7 @@ package com.xkcddatahub.fetcher.output.http.rest.xkcd.mappers
 
 import com.xkcddatahub.fetcher.adapters.output.http.rest.xkcd.data.ApiWebComics
 import com.xkcddatahub.fetcher.adapters.output.http.rest.xkcd.mappers.ApiWebComicsMapper.toDomain
-import com.xkcddatahub.fetcher.entity.WebComics
+import com.xkcddatahub.fetcher.domain.entity.WebComics
 import io.kotest.matchers.shouldBe
 import kotlin.test.Test
